\r to \n
