+++ as3-danmaku/host/Unpack/Button.ts
@@ -1,5 +1,5 @@
-import { createElement } from "../../worker/Utils";
 import { ScriptingContext } from "../ScriptingContext";
+import { createElement } from "../Utils";
 import { DisplayObject } from "./DisplayObject";
 
 export class Button extends DisplayObject {
